requirements.txt
fastapi
uvicorn[standard]
duckdb
pymongo
mysql-connector-python
typing